+++ src/client/js/index.js
@@ -4,6 +4,7 @@
 
 // import here templates required to display fields in Tabular
 import '../components/dt_email_address/dt_email_address.js';
+import '../components/dt_email_addresses/dt_email_addresses.js';
 import '../components/dt_email_more/dt_email_more.js';
 import '../components/dt_email_verified/dt_email_verified.js';
 //